Moves panel, master detail to samples/containers.
