+++ kotlin/samples/containers/android/build.gradle
@@ -1,0 +1,51 @@
+/*
+ * Copyright 2019 Square Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+apply plugin: 'com.android.library'
+apply plugin: 'kotlin-android'
+
+sourceCompatibility = JavaVersion.VERSION_1_8
+targetCompatibility = JavaVersion.VERSION_1_8
+
+android rootProject.ext.defaultAndroidConfig
+
+// See https://github.com/Kotlin/kotlinx.coroutines/issues/1064#issuecomment-479412940
+android.packagingOptions {
+  exclude 'META-INF/atomicfu.kotlin_module'
+  exclude 'META-INF/common.kotlin_module'
+}
+
+dependencies {
+  api project(':workflow-core')
+  api project(':workflow-ui-android')
+  api project(':samples:containers:common')
+
+  api deps.androidx.transition
+  api deps.kotlin.stdLib.jdk6
+  api deps.rxjava2.rxjava2
+
+  implementation project(':workflow-runtime')
+  implementation deps.androidx.appcompat
+  implementation deps.androidx.lifecycle.extensions
+  implementation deps.androidx.lifecycle.reactivestreams
+  implementation deps.androidx.savedstate
+  implementation deps.kotlin.coroutines.android
+  implementation deps.kotlin.coroutines.core
+  implementation deps.kotlin.coroutines.rx2
+
+  testImplementation deps.test.junit
+  testImplementation deps.test.truth
+  testImplementation deps.kotlin.coroutines.test
+}
