+++ kotlin/samples/containers/android/src/main/res/layout-ldrtl/master_detail_split.xml
