+++ kotlin/samples/containers/android/src/main/res/layout/master_detail_single.xml
