+++ kotlin/samples/containers/android/src/main/res/values-land/layout.xml
