+++ kotlin/samples/containers/android/src/main/res/values-sw600dp/bools.xml
