+++ kotlin/samples/containers/android/src/main/res/values/bools.xml
