+++ kotlin/samples/containers/android/src/main/res/values/ids.xml
@@ -16,5 +16,5 @@
   -->
 <resources>
   <!-- ID used by PanelContainer to opt in to view persistence. -->
-  <item type="id" name="tic_tac_workflow_panel_container"/>
+  <item type="id" name="panel_container"/>
 </resources>
