+++ kotlin/samples/containers/android/src/main/res/values/layout.xml
