+++ kotlin/samples/containers/common/src/main/java/com/squareup/sample/container/masterdetail/MasterDetailAware.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.squareup.sample.todo
+package com.squareup.sample.container.masterdetail
 
 /**
  * Informs [MasterDetailAware] renderings whether they're children of a [MasterDetailScreen],
