+++ kotlin/samples/dungeon/app/build.gradle
@@ -31,6 +31,7 @@ android {
   // See https://github.com/Kotlin/kotlinx.coroutines/issues/1064#issuecomment-479412940
   packagingOptions {
     exclude 'META-INF/atomicfu.kotlin_module'
+    exclude 'META-INF/common.kotlin_module'
   }
 }
