+++ kotlin/samples/hello-workflow/build.gradle
@@ -30,6 +30,7 @@ android {
   // See https://github.com/Kotlin/kotlinx.coroutines/issues/1064#issuecomment-479412940
   packagingOptions {
     exclude 'META-INF/atomicfu.kotlin_module'
+    exclude 'META-INF/common.kotlin_module'
   }
 }
