+++ kotlin/samples/tictactoe/app/build.gradle
@@ -33,12 +33,14 @@ android {
   // See https://github.com/Kotlin/kotlinx.coroutines/issues/1064#issuecomment-479412940
   packagingOptions {
     exclude 'META-INF/atomicfu.kotlin_module'
+    exclude 'META-INF/common.kotlin_module'
   }
 
   testOptions.unitTests.includeAndroidResources = true
 }
 
 dependencies {
+  implementation project(':samples:containers:android')
   implementation project(':samples:tictactoe:common')
   implementation project(':workflow-ui-android')
   implementation project(':workflow-core')
