+++ kotlin/samples/tictactoe/common/build.gradle
@@ -17,6 +17,7 @@ apply plugin: 'java'
 apply plugin: 'kotlin'
 
 dependencies {
+  implementation project(':samples:containers:common')
   implementation project(':workflow-ui-core')
   implementation project(':workflow-core')
   implementation project(':workflow-rx2')
