+++ kotlin/samples/tictactoe/common/src/test/java/com/squareup/sample/mainworkflow/MainWorkflowTest.kt
@@ -6,7 +6,7 @@ import com.squareup.sample.authworkflow.AuthWorkflow
 import com.squareup.sample.gameworkflow.GamePlayScreen
 import com.squareup.sample.gameworkflow.RunGameScreen
 import com.squareup.sample.gameworkflow.RunGameWorkflow
-import com.squareup.sample.panel.PanelContainerScreen
+import com.squareup.sample.container.panel.PanelContainerScreen
 import com.squareup.workflow.Worker
 import com.squareup.workflow.Workflow
 import com.squareup.workflow.WorkflowAction
@@ -53,7 +53,9 @@ class MainWorkflowTest {
 
   private fun runGameScreen(
     body: String = DEFAULT_RUN_GAME
-  ) = RunGameScreen(PanelContainerScreen(body))
+  ) = RunGameScreen(
+      PanelContainerScreen(body)
+  )
 
   private fun authScreen(wrapped: String = DEFAULT_AUTH) =
     BackStackScreen<Any>(wrapped)
