+++ kotlin/samples/todo-android/app/src/main/java/com/squareup/sample/mainactivity/MainActivity.kt
@@ -17,6 +17,7 @@ package com.squareup.sample.mainactivity
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.squareup.sample.container.masterdetail.MasterDetailContainer
 import com.squareup.sample.todo.TodoListsAppWorkflow
 import com.squareup.workflow.diagnostic.SimpleLoggingDiagnosticListener
 import com.squareup.workflow.diagnostic.andThen
@@ -44,6 +45,8 @@ class MainActivity : AppCompatActivity() {
 
   private companion object {
     val viewRegistry =
-      ViewRegistry(TodoEditorLayoutRunner, TodoListsLayoutRunner, MasterDetailContainer)
+      ViewRegistry(TodoEditorLayoutRunner, TodoListsLayoutRunner,
+          MasterDetailContainer
+      )
   }
 }
