+++ kotlin/settings.gradle
@@ -19,6 +19,8 @@ include ':benchmarks'
 include ':legacy:legacy-workflow-core'
 include ':legacy:legacy-workflow-rx2'
 include ':legacy:legacy-workflow-test'
+include ':samples:containers:android'
+include ':samples:containers:common'
 include ':samples:dungeon:app'
 include ':samples:dungeon:common'
 include ':samples:hello-terminal:hello-terminal-app'
